BUGFIX: the save/load functionnality was broken
The visibility had been turned into an enumeration.
When writting a enumeration element, it does write its name (e.g. "YES") and
not its integer value like in C (e.g. 0). Then the loading function was trying
to read an integer while it was a string.

The function has been lost since commit :
c83c86d73a24743e7d2e634a52048861250d7282

It was not elegant to use the displayed values in the JComboBox as the
informative value:
- we have removed the hard-coded strings that represent the values, and we
  have overidden the toString() method of EnumVisibility to show a more
  human-friendly ComboBox
- when saving the file, we want the visibility not to change accross versions,
  so we have print visibility.name() of the EnumVisibility vis

This should be added in the master but there is a commit historiy too large to
simply patch or pick out this individual commit. It will wait the next merging
with master.