Bugfix: fixed a bug that was already existing in Fizzim 14.02.26.
When we open several pages, and we create a state, and then want to
create a new Fizzim document, it crashed.

The fix is ugly (using side effects), but it is based on the initial
design that was working because of this very side effect. One day a
refactory on it will be required.