Bugfix: Duplicate names were not excluded since commit 3413e81
GeneralObj have 2 names: the one registered as field, and the one
registered as an attribute.

It was creating many lack of coherence. There is now only the attribute
field. The fix is not pretty since we need to send the global attributes
to the objects, but it will be ok for a while.